Minor version bump - 2.4.0  - Added support for custom header tags.  - Suppressed clang compiler warnings.
Apologies to all w/ outstanding issues for letting this project sit
idle, but updates will be coming out **much** more quickly now.

+++ src/utils/bamtools_fasta.cpp
@@ -77,7 +77,7 @@ void Fasta::FastaPrivate::Chomp(char* sequence) {
     static const int CHAR_LF = 10;
     static const int CHAR_CR = 13;
   
-    size_t seqLength = strlen(sequence);
+    int seqLength = strlen(sequence);
     if ( seqLength == 0 ) return;
     --seqLength; // ignore null terminator
